+++ src/extension.web.ts
@@ -104,7 +104,7 @@ import { initializeGlobals as initializeTelemetryGlobals } from './platform/tele
 import { setDisposableTracker } from './platform/common/utils/lifecycle';
 import { sendTelemetryEvent } from './telemetry';
 import { getVSCodeChannel } from './platform/common/application/applicationEnvironment';
-import { isUsingPylance } from './standalone/intellisense/notebookPythonPathService.node';
+import { isUsingPylance } from './standalone/intellisense/notebookPythonPathService';
 
 durations.codeLoadingTime = stopWatch.elapsedTime;
